fix: share snapshot_created fixture and fix its file path
Importing snapshot_created into test_local_snapshot_restoration.py
made pytest run it twice, once per module that requests it, since
pytest does not cache an imported fixture across the module that
defines it and the module that imports it. Moved it into conftest.py
so there is exactly one shared instance regardless of which test
file requests it first.

Separately, stage 2 of snapshot creation returned the snapshot file
name as-is from the script output, relative to cardano-db-sync (its
own cwd), but callers resolved it against their own cwd instead. The
restore step then looked for the file in the wrong directory. Stage 2
now returns the full path anchored to cardano-db-sync.
